Crate parity_local_store

Manages local node data: pending local transactions, sync security level

Structs

LocalDataStore

Manages local node data.

Enums

Error

Errors which can occur while using the local data store.

Traits

NodeInfo

Something which can provide information about the local node.

Functions

create

Create a new local data store, given a database, a column to write to, and a node. Attempts to read data out of the store, and move it into the node.